Change droid experience from float to uint32_t.
Less floats generally reduces the risk of desynchs, don't know if the droid experience caused desynchs, but it might have been possible. Added experience to syncDebugDroid.

Also, fixed QualityFactor and droidResistance integer truncation and made single-player skirmish damage not affected by difficulty level either.

+++ lib/netplay/nettypes.h
@@ -144,7 +144,7 @@ typedef struct PackagedCheck
 	int32_t order;
 	uint32_t secondaryOrder;
 	uint32_t body;
-	float experience;
+	uint32_t experience;
 	Position pos;
 	Rotation rot;
 	float sMoveX;
